feat: added the ability to provide another Compiler
